fix: cover the desktop hub and dedupe wecom/weixin's temp-dir path
Review of #2010 found the desktop hub (cmd/octo-desktop) never went
through cmd/octo's runServe, so it silently never ran the new uploads
housekeeping sweep — despite being the primary long-running "desktop
hub" scenario the issue calls out.

Moves the sweep itself into internal/server.StartUploadsHousekeeping
(both binaries can reach it without a cycle), leaving cmd/octo's
startUploadsHousekeeping as a thin wrapper that keeps its
housekeepingDisabled test guard, and wires the same call into
cmd/octo-desktop's startHub (safe unconditionally — that path only
runs from a live GUI event loop, never from `go test`).

Also routes wecom/weixin's already-namespaced temp dirs through the
new uploads.ChannelTempDir helper instead of inlining the same path
construction twice, so there's one source of truth for what the sweep
targets.
