+++ internal/server/attachments.go
@@ -93,12 +93,6 @@ type userAttachments struct {
 	notes []string
 }
 
-// UploadsDir returns ~/.octo/uploads, creating it if needed. Exported for
-// cmd/octo's startup housekeeping sweep (internal/uploads.Sweep).
-func UploadsDir() (string, error) {
-	return ensureUploadsDir()
-}
-
 // ensureUploadsDir returns ~/.octo/uploads, creating it if needed.
 func ensureUploadsDir() (string, error) {
 	home, err := os.UserHomeDir()
